Major refactoring, additions, and improvements
-Refactored a lot of code
-Added duo-confirmation functionality for transactions
-Added emails to confirmations/cancellations
-Added custom routes to make our urls prettier
-Added new favicon
-Improved layouts of some pages
-Improved flow of matching posts and confirming/cancelling transactions
-Imported Bootstrap 3 to use for Home page layout (not yet implemented)

+++ BookSpade/BookSpade.Revamped/Models/Post.cs
@@ -12,7 +12,7 @@ namespace BookSpade.Revamped.Models
         public int UserId { get; set; }
         public int TextBookId { get; set; }
         public ActionBy ActionBy { get; set; }
-        public decimal Price { get; set; }
+        public int Price { get; set; }
         public BookCondition BookCondition { get; set; }
         public int IsTransacting { get; set; }
         public int IsActive { get; set; }
@@ -25,7 +25,7 @@ namespace BookSpade.Revamped.Models
                 int userId,
                 int textBookId,
                 ActionBy actionBy,
-                decimal price,
+                int price,
                 BookCondition bookCondition,
                 int isTransacting,
                 int isActive,
